% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DErsql-package.R
\docType{package}
\name{DErsql-package}
\alias{DErsql}
\alias{DErsql-package}
\title{DErsql: Functions for DE}
\description{
Internal package for helping with workflow.
}
\author{
\strong{Maintainer}: Allan Butler \email{allanbutler9@gmail.com}

}
\keyword{internal}
