% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{ImportEvergreenData}
\alias{ImportEvergreenData}
\title{ImportEvergreenData}
\usage{
ImportEvergreenData(LatestRatePlanGroup = "Evergreen",
  StartTime = "2018-01-01", EndTime = "2018-12-01")
}
\description{
This function allows you import Evergreen customer data from the ODS.ODS_RO.vw_us_account_base.
}
\examples{
ImportEvergreenData()
}
